Update alt prive & credits subscription tests

+++ app/costasiella/schema/account_subscription_alt_price.py
@@ -78,7 +78,7 @@ class CreateAccountSubscriptionAltPrice(graphene.relay.ClientIDMutation):
         account_subscription = graphene.ID(required=True)
         subscription_year = graphene.Int(required=True)
         subscription_month = graphene.Int(required=True)
-        amount = graphene.Float(required=False, default_value=0)
+        amount = graphene.Decimal(required=False, default_value=0)
         description = graphene.String(required=False, default_value="")
         note = graphene.String(required=False, default_value="")
 
@@ -113,7 +113,7 @@ class UpdateAccountSubscriptionAltPrice(graphene.relay.ClientIDMutation):
         id = graphene.ID(required=True)
         subscription_year = graphene.Int(required=True)
         subscription_month = graphene.Int(required=True)
-        amount = graphene.Float(required=False, default_value=0)
+        amount = graphene.Decimal(required=False, default_value=0)
         description = graphene.String(required=False, default_value="")
         note = graphene.String(required=False, default_value="")
